Major change to the use of SSHCredentials
SSH keys are now stored inside the keystore as part of SSH Credentials. Previously only the path to the key file was stored. This key is now sent from the client to the SaaS node when deploying a processor. Previously, the identity that is running the node was expected to have the required SSH Credentials and only the name of the SSH profile was passed from the client. This unnecessary inflexible and doesn't work well with the CLI (or any other client for that matter). SSH keys are now stored on disk only for the time of deployment (i.e., they are delete when a processor is undeployed again).