Use snappy for file compression
Note: the content of the file still has to be read into the memory but the total amount of data that needs to be transferred is reduced. Compression also resolves the issue with The Weird TIFF file (which was probably due to some encoding issues somewhere deep down in the library code..).